requirements.txt
llama-index
llama-index-llms-google-genai
pypdf
python-dotenv
Ipython
streamlit
LLama-index-llms-gemini
google-generativeai
